webapi/webapi-rawsockets-w3c-tests/rawsockets/UDPSocket_event.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Shentu.Jiazhen <jiazhenx.shentu@intel.com>

-->

<meta charset='utf-8'>
<title>Raw Socket Test: UDPSocket event</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/2012/sysapps/raw-sockets/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/support.js"></script>
<div id="log"></div>
<script>
// These cases still under designed status.
test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  socket.addEventListener("drain", handleDrain);
  function handleDrain (event) { }
  assert_unreached("TODO debug");
}, "Check if set up the ondrain event successful for UDPSocket using the type drain");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  socket.addEventListener("error", handleError);
  function handleError (event) { }
  assert_unreached("TODO debug");
}, "Check if set up the onerror event successful for UDPSocket using the type error");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  socket.addEventListener("message", handleMessage);
  function handleMessage (event) { }
  assert_unreached("TODO debug");
}, "Check if set up the onmessage event successful for UDPSocket using the type message");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  socket.addEventListener("open", handleOpen);
  function handleOpen (event) { }
  assert_unreached("TODO debug");
}, "Check if set up the onopen event successful for UDPSocket using the type open");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  socket.ondrain = function(){ };
  assert_unreached("TODO debug");
}, "Check if ondrain event can be fired upon detection that previously-buffered data has been written to the network and it is possible to buffer more data received from the application");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  socket.onerror = function(){ };
  assert_unreached("TODO debug");
}, "Check if onerror event can be fired when there is an error");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  socket.onmessage = function(){ };
  assert_unreached("TODO debug");
}, "Check if onmessage event can be fired when every time a UDP datagram has been received and was read");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  socket.onopen = function(){ };
  assert_unreached("TODO debug");
}, "Check if onopen event can be fired when the socket is open and ready to use to send and receive data");

</script>
